requirements.txt
Django
djangorestframework
pytest
markdown
django-filter
django-bootstrap4
mysqlclient
stripe
django-cors-headers
